+++ lib/views/view_aliases/view_aliases.dart
@@ -55,7 +55,7 @@ class ViewAliasesState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {