+++ lib/views/view_categories/view_categories.dart
@@ -167,7 +167,7 @@ class ViewCategoriesState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {