+++ lib/views/view_loans/view_loans.dart
@@ -49,7 +49,7 @@ class ViewLoansState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {