+++ lib/views/view_payees/view_payees.dart
@@ -57,7 +57,7 @@ class ViewPayeesState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {