+++ lib/views/view_rentals/view_rentals.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:money/helpers/list_helper.dart';
-import 'package:money/models/fields/fields.dart';
+import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/storage/data/data.dart';
-import 'package:money/models/money_objects/money_object.dart';
-
 import 'package:money/models/money_objects/rent_buildings/rent_building.dart';
 import 'package:money/models/money_objects/rental_unit/rental_unit.dart';
 import 'package:money/models/money_objects/splits/split.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
 import 'package:money/views/view_rentals/rental_pnl.dart';
 import 'package:money/views/view_rentals/rental_pnl_card.dart';
-import 'package:money/widgets/center_message.dart';
-
 import 'package:money/widgets/chart.dart';
 import 'package:money/views/view_money_objects.dart';
 import 'package:money/widgets/list_view/transactions/list_view_transactions.dart';
@@ -26,6 +22,12 @@ class ViewRentals extends ViewForMoneyObjects {
 }
 
 class ViewRentalsState extends ViewForMoneyObjectsState {
+  RentBuilding? lastSelectedRental;
+
+  ViewRentalsState() {
+    onCopyInfoPanelTransactions = _onCopyInfoPanelTransactions;
+  }
+
   @override
   String getClassNamePlural() {
     return 'Rentals';
@@ -64,11 +66,16 @@ class ViewRentalsState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {
     return _getSubViewContentForTransactions(selectedIds);
+  }
+
+  void _onCopyInfoPanelTransactions() {
+    final list = getTransactionLastSelectedItem();
+    copyToClipboardAndInformUser(context, MoneyObjects.getCsvFromList(list));
   }
 
   @override