+++ lib/views/view_stocks/view_stocks.dart
@@ -19,6 +19,12 @@ class ViewStocks extends ViewForMoneyObjects {
 }
 
 class ViewStocksState extends ViewForMoneyObjectsState {
+  Security? lastSecuritySelected;
+
+  ViewStocksState() {
+    onCopyInfoPanelTransactions = _onCopyInfoPanelTransactions;
+  }
+
   @override
   String getClassNamePlural() {
     return 'Stocks';
@@ -64,29 +70,40 @@ class ViewStocksState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {
-    final Security? security = getMoneyObjectFromFirstSelectedId<Security>(selectedIds, getList());
-    if (security == null) {
+    lastSecuritySelected = getMoneyObjectFromFirstSelectedId<Security>(selectedIds, getList());
+    if (lastSecuritySelected == null) {
       return const CenterMessage(message: 'No security selected.');
     }
 
-    final List<Investment> list = Investments.getInvestmentsFromSecurity(security.uniqueId);
-    Investments.calculateRunningBalance(list);
+    List<Investment> list = getListOfInvestment(lastSecuritySelected!);
 
     final exclude = ['Symbol', 'Load', 'Fees'];
     final fieldsToDisplay = Investment.fields.definitions
         .where((element) => element.useAsColumn && !exclude.contains(element.name))
         .toList();
+
     return AdaptableListView(
-        fieldDefinitions: fieldsToDisplay,
-        list: list,
-        sortAscending: false,
-        selectedItemsByUniqueId: ValueNotifier<List<int>>(<int>[]),
-        onItemTap: (BuildContext __, int _) {
-          copyToClipboardAndInformUser(context, MoneyObjects.getCsvFromList(list));
-        });
+      fieldDefinitions: fieldsToDisplay,
+      list: list,
+      sortAscending: false,
+      selectedItemsByUniqueId: ValueNotifier<List<int>>(<int>[]),
+    );
+  }
+
+  List<Investment> getListOfInvestment(Security security) {
+    final List<Investment> list = Investments.getInvestmentsFromSecurity(security.uniqueId);
+    Investments.calculateRunningBalance(list);
+    return list;
+  }
+
+  void _onCopyInfoPanelTransactions() {
+    if (lastSecuritySelected != null) {
+      final list = getListOfInvestment(lastSecuritySelected!);
+      copyToClipboardAndInformUser(context, MoneyObjects.getCsvFromList(list));
+    }
   }
 }