+++ lib/views/view_transactions/view_transactions.dart
@@ -185,7 +185,7 @@ class ViewTransactionsState extends ViewForMoneyObjectsState {
   }
 
   @override
-  Widget getPanelForTransactions({
+  Widget getInfoPanelSubViewTransactions({
     required final List<int> selectedIds,
     required final bool showAsNativeCurrency,
   }) {