+++ lib/widgets/details_panel/info_panel_views_enum.dart
@@ -1,4 +1,4 @@
-enum SubViewsEnum {
+enum InfoPanelSubViewEnum {
   details,
   chart,
   transactions,